+++ src/zobrist.h
@@ -23,7 +23,7 @@ constexpr uint64_t random_uint64_t() {
 	return seed;
 }
 constexpr uint64_t hash_pce(uint64_t zobrist_key, Piece pce, const int sq) {
-	return zobrist_key ^ keys_pce[static_cast<int>(pce)][sq];
+	return zobrist_key ^ keys_pce[pce][sq];
 }
 constexpr uint64_t hash_castling_rights(uint64_t zobrist_key, const int castling_rights) {
 	return zobrist_key ^ keys_castling_rights[castling_rights];
@@ -32,5 +32,5 @@ constexpr uint64_t hash_side_to_move(uint64_t zobrist_key) {
 	return zobrist_key ^ key_side_to_move;
 }
 inline uint64_t hash_en_passant_sq(uint64_t zobrist_key, Square sq) {
-	return zobrist_key ^ keys_pce[static_cast<int>(Piece::NONE)][static_cast<int>(sq)];
+	return zobrist_key ^ keys_pce[Piece::NONE][sq];
 }
